Add command creation test
