+++ engine/launcher/src/test/scala/org/enso/launcher/components/ComponentsManagerSpec.scala
@@ -1,53 +1,9 @@
 package org.enso.launcher.components
 
-import java.nio.file.Path
-
 import nl.gn0s1s.bump.SemVer
-import org.enso.launcher.cli.GlobalCLIOptions
-import org.enso.launcher.installation.DistributionManager
-import org.enso.launcher.releases.{
-  EngineReleaseProvider,
-  GraalCEReleaseProvider
-}
-import org.enso.launcher.{FakeEnvironment, Logger, WithTemporaryDirectory}
-import org.scalatest.OptionValues
-import org.scalatest.matchers.should.Matchers
-import org.scalatest.wordspec.AnyWordSpec
+import org.enso.launcher.Logger
 
-class ComponentsManagerSpec
-    extends AnyWordSpec
-    with Matchers
-    with OptionValues
-    with WithTemporaryDirectory
-    with FakeEnvironment {
-
-  private def makeManagers(): (DistributionManager, ComponentsManager) = {
-    val distributionManager = new DistributionManager(
-      fakeInstalledEnvironment()
-    )
-    val fakeReleasesRoot =
-      Path.of(
-        getClass
-          .getResource("fake-releases")
-          .toURI
-      )
-    val engineProvider = new EngineReleaseProvider(
-      FakeReleaseProvider(fakeReleasesRoot.resolve("enso"))
-    )
-    val runtimeProvider = new GraalCEReleaseProvider(
-      FakeReleaseProvider(fakeReleasesRoot.resolve("graalvm"))
-    )
-    val componentsManager = new ComponentsManager(
-      GlobalCLIOptions(autoConfirm = true, hideProgress = true),
-      distributionManager,
-      engineProvider,
-      runtimeProvider
-    )
-
-    (distributionManager, componentsManager)
-  }
-
-  def makeComponentsManager(): ComponentsManager = makeManagers()._2
+class ComponentsManagerSpec extends ComponentsManagerTest {
 
   "ComponentsManager" should {
     "find the latest engine version in semver ordering" in {
